Changed ingredient amount column to float

+++ Dtos/IngredientDtos/GetIngredientDto.cs
@@ -5,5 +5,5 @@ public class GetIngredientDto
     public int Id { get; set; }
     public string Name { get; set; }
     public string Unit { get; set; }
-    public int Amount { get; set; }
+    public float Amount { get; set; }
 }
